+++ ChangeLog.md
@@ -42,6 +42,9 @@ authentication.
 configuration file stored in ~/.ssh/config (or the location specified in the
 `SSHConfig` parameter), if the file exists.  Parameters read from the OpenSSH
 configuration file will take precedence over any TurboVNC Viewer parameters.
+     - A new Java system property (`turbovnc.sshauth`) can be used to enable or
+disable SSH authentication methods, as well as to specify their preferred
+order.
 
 6. Fixed a race condition in the TurboVNC Server that, under rare
 circumstances, caused the TurboVNC Viewer to incorrectly assume that the server